Method lsei definition corrected.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/limsolve@119 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/limSolve/src/limSolveMethods.c
@@ -5,7 +5,7 @@
 
 
 void F77_NAME(ldp)(double*, double*, int*, int*, int*,  double*, 
-      double*, double*, double*, int*, int*, int*, int*);
+      double*, double*, int*, int*, int*, int*, int*);
      
 void F77_NAME(lsei)(int*, int*, int*, int*, double*, double*,
       double*, double*, double*, double*, double*, int*, int*,
